moves path handling logic for ViewChargers into ViewChargers component
