+++ src/layouts/Navbar.js
@@ -1,14 +1,12 @@
 import { React, useEffect } from "react";
 import { Toolbar, AppBar, useMediaQuery, useTheme } from "@mui/material";
 import { useLocation } from "react-router-dom";
-import { useGlobalState } from "../context/stateContext";
-import { getMyChargers, getChargers } from "../services/chargerServices";
 
+import { useGlobalState } from "../context/stateContext";
 import { MobileNavbar } from "./MobileNavbar";
 import { DesktopNavBar } from "./DesktopNavBar";
 
 export const Navbar = () => {
-  /** For access to MUI Breakpoints */
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down("md"));
   const location = useLocation();
@@ -19,9 +17,7 @@ export const Navbar = () => {
       type: "setLocation",
       data: location,
     });
-
-    fetchData(location, dispatch);
-  }, [location, dispatch]);
+  }, [location]);
 
   return (
     <>
@@ -41,33 +37,3 @@ export const Navbar = () => {
     </>
   );
 };
-
-async function fetchData(location, dispatch) {
-  if (location.pathname === "/chargers/mychargers") {
-    try {
-      const myChargers = await getMyChargers();
-      if (myChargers) {
-        dispatch({
-          type: "setChargerList",
-          data: myChargers,
-        });
-      } else {
-        console.log("There's an error in fetching myChargers");
-      }
-    } catch (err) {
-      console.log(err.message);
-    }
-  } else if (location.pathname === "/chargers") {
-    try {
-      const chargers = await getChargers();
-      if (chargers) {
-        dispatch({
-          type: "setChargerList",
-          data: chargers,
-        });
-      }
-    } catch (err) {
-      console.log(err.message);
-    }
-  }
-}
