+++ src/services/chargerServices.js
@@ -65,9 +65,7 @@ export async function getChargers() {
     // TODO: Check what comes back from the data to handle messages
     return response.data;
   } catch (err) {
-    // console.log("Axios ERROR----");
-    // console.log(err.message);
-    return err.response;
+    throw Error(err.response.data.error);
   }
 }
 
@@ -78,8 +76,6 @@ export async function getMyChargers() {
     // TODO: Check what comes back from the data to handle messages
     return response.data;
   } catch (err) {
-    // console.log("Axios ERROR----");
-    // console.log(err.message);
-    return err.response;
+    throw Error(err.response.data.error);
   }
 }
